Scaffolding out control methods for tag manager
