+++ lib/modules/apostrophe-tags/lib/browser.js
@@ -2,6 +2,7 @@ var _ = require('lodash');
 
 module.exports = function(self, options) {
   self.pushAssets = function() {
+    self.pushAsset('stylesheet', 'user', { when: 'user' });
     self.pushAsset('script', 'user', { when: 'user' });
     self.pushAsset('script', 'manager-modal', { when: 'user' });
   };
