+++ lib/modules/apostrophe-tags/public/css/user.less
@@ -1,0 +1,28 @@
+.apos-tags-view
+{
+  .apos-manage-tags
+  {
+    margin: 0 auto;
+    width: 50%;
+    min-width: 400px;
+  }
+  .apos-manage-tag
+  {
+    position: relative;
+  }
+  .apos-manage-tag-label .apos-field-input:disabled
+  {
+    padding: @apos-padding-1 @apos-padding-2;
+  }
+  .apos-manage-tag-controls
+  {
+    top: 50%;
+    transform: translateY(-50%);
+    position: absolute;
+    right: 0;
+  }
+  .apos-manage-tag-control
+  {
+    padding: @apos-padding-1;
+  }
+}
