+++ lib/modules/apostrophe-tags/public/js/manager-modal.js
@@ -4,7 +4,6 @@ apos.define('apostrophe-tags-manager-modal', {
   construct: function(self, options) {
     self.beforeShow = function(callback) {
       self.$results = self.$el.find('[data-apos-tags-view]');
-      self.$pager = self.$el.find('[data-apos-pager]');
       return self.refresh(callback);
     };
 
@@ -18,16 +17,26 @@ apos.define('apostrophe-tags-manager-modal', {
           return;
         }
         self.$results.html(response.data.results);
-        self.$pager.html(response.data.pager);
         apos.emit('enhance', self.$results);
-        apos.emit('enhance', self.$pager);
         self.resizeContentHeight();
         self.afterRefresh();
 
         if (callback) {
-          return callback(response);
+          return callback();
         }
       });
+    };
+
+    self.add = function() {
+
+    };
+
+    self.edit = function() {
+
+    };
+
+    self.trash = function() {
+
     };
 
     self.beforeRefresh = function(options) {
